plugins: split package IO from provider build internals
Interfaces:\n- services/plugins/packageio now owns manifest, archive, static catalog, reference, digest, platform, and UI-route primitives. Example: packageio.ReadPackageManifest(path), packageio.ExtractPackage(path, root), packageio.ReadStaticCatalog(root, name).\n- services/plugins.ValidationPluginFromManifest remains config-free and wires packageio static catalog loading plus plugins.DefaultAPICatalogLoader. Example: plugins.ValidateEffectiveCatalog(ctx, name, plugins.ValidationPluginFromManifest(manifestPath, manifest)).\n- internal/config.PluginValidationConfig adapts daemon config.ProviderEntry values into plugins.ValidationConfig. Example: plugins.ValidateEffectiveCatalogsAndDependencies(ctx, config.PluginValidationConfig(cfg)).\n\nservices/plugins/providerpkg remains a compatibility facade for existing provider packaging callers while source build/release internals stay there.
